requirements.txt
bcrypt==4.2.1
blinker==1.6.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.6
MarkupSafe==2.1.5
phonenumbers==9.0.7
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2025.2
requests==2.31.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
